Manuel_Figueroa/requirements.txt
pycryptodome==3.23.0
pytest==9.0.2
typer==0.24.1
rich==14.3.3
ruff==0.15.5
